backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
motor==3.6.0
pydantic==2.9.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
httpx==0.27.2
